+++ src/main/java/com/example/springbatch/DBJobConfiguration.java
@@ -1,0 +1,68 @@
+/*******************************************************************************
+ * created by        : jws
+ * creation date     : 2022-01-03
+ *
+ * Copyright (c) 2021 Samsung SDS.
+ * All rights reserved.
+ *******************************************************************************/
+
+package com.example.springbatch;
+
+
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.StepContribution;
+import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
+import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
+import org.springframework.batch.core.scope.context.ChunkContext;
+import org.springframework.batch.core.step.tasklet.Tasklet;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import lombok.RequiredArgsConstructor;
+
+
+/**
+ * Class description
+ *
+ * @author jws
+ * @since 2022. 1. 3
+ * @version 1.0
+*/
+
+@RequiredArgsConstructor
+@Configuration
+public class DBJobConfiguration {
+
+	private final JobBuilderFactory		jobBuilderFactory;
+	private final StepBuilderFactory	stepBuilderFactory;
+
+	@Bean
+	public Job job() {
+		return jobBuilderFactory.get("job")
+				.start(step1())
+				.next(step2())
+				.build();
+	}
+
+	public Step step2() {
+		return stepBuilderFactory.get("step2").tasklet(new Tasklet() {
+			@Override
+			public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
+				return RepeatStatus.FINISHED;
+			}
+		})
+		.build();
+	}
+
+	public Step step1() {
+		return stepBuilderFactory.get("step1").tasklet(new Tasklet() {
+			@Override
+			public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
+				return RepeatStatus.FINISHED;
+			}
+		})
+		.build();
+	}
+}
