+++ src/main/resources/application-dev.yml
@@ -1,0 +1,8 @@
+spring:
+  config:
+    activate:
+      on-profile: dev
+  datasource:    hikari:      jdbc-url: 
+      username: sa
+      password: 
+      driver-class-name: org.h2.Driver
